+++ src/EjercicioZoologico/PavoReal.java
@@ -1,4 +1,4 @@
-package Zoologico;
+package EjercicioZoologico;
 
 public class PavoReal extends Aviario{
     @Override
